feat(frontend): bulk tool selection and policy actions on the server setup Tools & Guardrails step
Per-tool checkboxes with select-all (scoped to the active name filter) and
the same bulk actions bar the guardrails table has: apply a call policy or
result policy to the selection, or configure the selection with a subagent.
The search filter sits above the bar as its own control.

The bar (desktop + mobile layouts, mutation state, custom-policy skip
logic) is extracted from the guardrails table into the shared
ToolPolicyBulkActionsBar component and reused by both surfaces; the
wizard's old all-tools Configure with Subagent button is superseded by the
selection-scoped one in the bar.
